requirements.txt
pqcow-liboqs==0.10.1.dev0
nose2==0.15.1
python-gnupg==0.5.3
cffi==1.17.1 
cryptography==44.0.0 
pycparser==2.22